+++ gui/canvas.go
@@ -1,4 +1,4 @@
-package fynerisor
+package gui
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 
 	"fyne.io/fyne/v2/canvas"
 
-	risorcanvas "github.com/uidbz/fynerisor/canvas"
+	risorcanvas "github.com/uidbz/fynerisor/gui/canvas"
 	"github.com/deepnoodle-ai/risor/v2/pkg/object"
 	"github.com/deepnoodle-ai/risor/v2/pkg/op"
 )