+++ gui/fyne.go
@@ -1,4 +1,4 @@
-package fynerisor
+package gui
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 
 	"fyne.io/fyne/v2"
 
-	risorwidget "github.com/uidbz/fynerisor/widget"
+	risorwidget "github.com/uidbz/fynerisor/gui/widget"
 	"github.com/deepnoodle-ai/risor/v2/pkg/object"
 	"github.com/deepnoodle-ai/risor/v2/pkg/op"
 )