+++ gui/widget/table.go
@@ -9,7 +9,7 @@ import (
 	"github.com/deepnoodle-ai/risor/v2/pkg/object"
 	"github.com/deepnoodle-ai/risor/v2/pkg/op"
 
-	"github.com/uidbz/fynerisor/widget/tablewidget"
+	"github.com/uidbz/fynerisor/gui/widget/tablewidget"
 )
 
 const TableType object.Type = "widget.Table"